minor code reorg for bound property controls
