+++ app/controls/ElasticPropertyValueBase.cpp
@@ -46,72 +46,4 @@ namespace ctb::app
    }
 
 
-   auto PropertyValueCtrl::create(wxWindow* parent, const DatasetEventSourcePtr& source, CtProp bound_prop) -> PropertyValueCtrl*
-   {
-      return detail::createDatasetWindow<PropertyValueCtrl>(parent, source, bound_prop);
-   }
-
-   // clang-format off
-   auto PropertyValueCtrl::create(wxWindow* parent, const DatasetEventSourcePtr& source, CtProp bound_prop,
-                                  std::string_view format, std::string_view null_display) -> PropertyValueCtrl*
-   {
-      auto* ctrl = detail::createDatasetWindow<PropertyValueCtrl>(parent, source, bound_prop);
-
-      ctrl->setNullDisplayValue(null_display);
-      if (!format.empty())
-      {
-         ctrl->setFormat(format);
-      }
-      return ctrl;
-   }
-   // clang-format on
-
-   auto PropertyValueCtrl::getDisplayValue(const DatasetPtr& ds, int rec_idx) const -> std::string
-   {
-      auto val = ds->getProperty(rec_idx, m_prop);
-
-      if (val.isNull()) return m_null_display;
-
-      return val.asString(m_format_str);
-   }
-
-
-   auto DrinkWindowCtrl::create(wxWindow* parent, const DatasetEventSourcePtr& source, CtProp begin_prop, CtProp end_prop) -> DrinkWindowCtrl*
-   {
-      return detail::createDatasetWindow<DrinkWindowCtrl>(parent, source, begin_prop, end_prop);
-   }
-
-
-   auto DrinkWindowCtrl::getDisplayValue(const DatasetPtr& ds, int rec_idx) const -> std::string
-   {
-      std::string value{};
-
-      if (ds->hasProperty(m_end_prop))
-      {
-         auto begin_dt = ds->getProperty(rec_idx, m_begin_prop);
-         auto end_dt   = ds->getProperty(rec_idx, m_end_prop);
-         value         = ctb::detail::getDrinkWindow(begin_dt, end_dt);
-      }
-      return value;
-   }
-
-
-   auto ProReviewsCacheCtrl::create(wxWindow* parent, const DatasetEventSourcePtr& source, CacheValueFn value_fn) -> ProReviewsCacheCtrl*
-   {
-      return detail::createDatasetWindow<ProReviewsCacheCtrl>(parent, source, value_fn);
-   }
-
-   auto ProReviewsCacheCtrl::getDisplayValue(const DatasetPtr& ds, int rec_idx) const -> std::string
-   {
-      std::string value{};
-
-      auto cache = wxGetApp().getProReviewsCache();
-      if (cache)
-      {
-         auto wine_id = ds->getProperty(rec_idx, CtProp::iWineId).asUInt64().value_or(0);
-         value        = (*cache.*m_value_fn)(wine_id);
-      }
-      return value;
-   }
-
 }   // namespace ctb::app
