+++ app/controls/ElasticPropertyValueBase.h
@@ -42,100 +42,8 @@ namespace ctb::app
    };
 
 
-   /// @brief standard/default property control that displays a single property
-   ///
-   class PropertyValueCtrl : public ElasticPropertyValueBase
-   {
-   public:
-      // clang-format off
-      /// @brief static factory method for creating PropertyValueCtrl objects
-      /// @param parent parent window for the control, must be non-null
-      /// @param source dataset event source to bind the control to, must be non-null
-      /// @param bound_prop the property id to bind the control to
-      /// @return non-owning pointer to the newly created window.
-      [[nodiscard]] static auto create(wxWindow* parent, const DatasetEventSourcePtr& source, CtProp bound_prop) -> PropertyValueCtrl*;
-      [[nodiscard]] static auto create(wxWindow* parent, const DatasetEventSourcePtr& source, CtProp bound_prop,
-                                       std::string_view format, std::string_view null_display) -> PropertyValueCtrl*;
-      // clang-format on
-
-      /// @brief Set the display format.
-      ///
-      /// Default is "{}" which just displays the string property, but you can change it if needed (e.g. currency etc)
-      auto setFormat(std::string_view fmt_str)
-      {
-         m_format_str = fmt_str;
-      }
-
-      /// @brief Set the value to display when the bound field isNull(). Default is empty string
-      auto setNullDisplayValue(std::string_view val)
-      {
-         m_null_display = val;
-      }
-
-   private:
-      CtProp      m_prop{};
-      std::string m_format_str{ constants::FMT_DEFAULT_FORMAT };
-      std::string m_null_display{};
-
-      DECLARE_DATASET_WINDOW_FACTORY;
-
-      PropertyValueCtrl(const DatasetEventSourcePtr& source, CtProp bound_prop) : ElasticPropertyValueBase{ source }, m_prop{ bound_prop }
-      {}
-
-      auto getDisplayValue(const DatasetPtr& ds, int rec_idx) const -> std::string override;
-   };
 
 
-   /// @brief This class uses drink start/end properties to display a formatted drinking window.
-   ///
-   class DrinkWindowCtrl : public ElasticPropertyValueBase
-   {
-   public:
-      /// @brief static factory method for creating DrinkWindowCtrl objects
-      /// @param parent parent window for the control, must be non-null
-      /// @param source dataset event source to bind the control to, must be non-null
-      /// @param bound_prop the property id to bind the control to
-      /// @return non-owning pointer to the newly created window.
-      [[nodiscard]] static auto create(wxWindow* parent, const DatasetEventSourcePtr& source, CtProp begin_prop, CtProp end_prop) -> DrinkWindowCtrl*;
-
-   private:
-      CtProp m_begin_prop{};
-      CtProp m_end_prop{};
-
-      DrinkWindowCtrl(const DatasetEventSourcePtr& source, CtProp begin_prop, CtProp end_prop)
-         : ElasticPropertyValueBase{ source }, m_begin_prop{ begin_prop }, m_end_prop{ end_prop }
-      {}
-
-      auto getDisplayValue(const DatasetPtr& ds, int rec_idx) const -> std::string override;
-
-      DECLARE_DATASET_WINDOW_FACTORY;
-   };
-
-
-   /// @brief This class binds a property value control to one of the getXXX() methods in the ProReviewsCache class.
-   ///
-   class ProReviewsCacheCtrl final : public ElasticPropertyValueBase
-   {
-   public:
-      using CacheValueFn = std::string (ProReviewsCache::*)(uint64_t) const;
-
-      /// @brief static factory method for creating ProReviewsCacheCtrl objects
-      /// @param parent the parent window for the control, must be non-null
-      /// @param source the dataset event source to bind the control to, must be non-null
-      /// @param bound_prop the property id to bind the control to
-      /// @return non-owning pointer to the newly created window.
-      [[nodiscard]] static auto create(wxWindow* parent, const DatasetEventSourcePtr& source, CacheValueFn value_fn) -> ProReviewsCacheCtrl*;
-
-   private:
-      CacheValueFn m_value_fn{};
-
-      ProReviewsCacheCtrl(const DatasetEventSourcePtr& source, CacheValueFn value_fn) : ElasticPropertyValueBase{ source }, m_value_fn{ std::move(value_fn) }
-      {}
-
-      auto getDisplayValue(const DatasetPtr& ds, int rec_idx) const -> std::string override;
-
-      DECLARE_DATASET_WINDOW_FACTORY;
-   };
 
 
 }   // namespace ctb::app
